# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-api-controller
      operationId: getSpecificUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: USER not found
        "200":
          description: USER ID CORRECT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "400":
          description: Invalid id supplied
    put:
      tags:
      - user-api-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: USER PUT CORRECT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "404":
          description: USER not found
        "401":
          description: Unauthorized
        "400":
          description: Invalid id supplied
    delete:
      tags:
      - user-api-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: USER DELETE CORRECT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "404":
          description: USER not found
        "401":
          description: Unauthorized
        "400":
          description: Invalid id supplied
  /api/users/{id}/image:
    get:
      tags:
      - user-api-controller
      operationId: getUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "401":
          description: Unauthorized
        "404":
          description: USER image not found
        "200":
          description: USER ID CORRECT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "400":
          description: Invalid id supplied
    put:
      tags:
      - user-api-controller
      operationId: updateUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - profileImage
              type: object
              properties:
                profileImage:
                  type: string
                  format: binary
      responses:
        "404":
          description: USER not found
        "200":
          description: USER IMAGE PUT CORRECT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "401":
          description: Unauthorized
        "400":
          description: Invalid id supplied
    post:
      tags:
      - user-api-controller
      operationId: uploadUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              required:
              - profileImage
              type: object
              properties:
                profileImage:
                  type: string
                  format: binary
      responses:
        "201":
          description: USER IMAGE POST CORRECT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "404":
          description: USER not found
        "401":
          description: Unauthorized
        "400":
          description: Invalid id supplied
    delete:
      tags:
      - user-api-controller
      operationId: deleteUserImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: USER not found
        "204":
          description: USER IMAGE DELETE CORRECT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Element'
        "401":
          description: Unauthorized
        "400":
          description: Invalid id supplied
...
  /api/users/:
    get:
      tags:
      - user-api-controller
      operationId: getUsers
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageUser'
    post:
      tags:
      - user-api-controller
      operationId: addUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    UserDTO:
      type: object
      properties:
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        password:
          type: string
    Element:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        author:
          type: string
        year:
          type: integer
          format: int32
        type:
          type: string
        season:
          type: string
        state:
          type: string
        country:
          type: string
        base64Image:
          type: string
        generos:
          type: array
          items:
            type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageUser:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/User'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        roles:
          type: array
          items:
            type: string
        elements:
          type: array
          items:
            $ref: '#/components/schemas/Element'
...